a2a-simple/pyproject.toml
[project]
name = "a2a-simple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.3.3",
    "sse-starlette>=3.0.2",
    "starlette>=0.47.3",
    "uvicorn>=0.35.0",
]
